refactor home components
