+++ Client/index.html
@@ -3,8 +3,9 @@
   <head>
     <meta charset="UTF-8" />
     <link rel="icon" type="image/svg+xml" href="/vite.svg" />
+    <meta name="description" content="Trung Tâm Thể Dục Thể Thao Số 1 Việt Nam" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Vite + React</title>
+    <title>Trung Tâm Thể Dục Thể Thao Số 1 Việt Nam</title>
   </head>
   <body>
     <div id="root"></div>
